% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{autocompletar}
\alias{autocompletar}
\title{Autocompletar palabra}
\usage{
autocompletar(texto)
}
\arguments{
\item{texto}{El texto parcial para buscar sugerencias.}
}
\value{
Un vector de caracteres con las sugerencias.
}
\description{
Autocompletar palabra
}
\examples{
\donttest{
autocompletar("progra")
}
}
