% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{buscar_anagrama}
\alias{buscar_anagrama}
\title{Buscar anagramas}
\usage{
buscar_anagrama(palabra)
}
\arguments{
\item{palabra}{La palabra para buscar anagramas.}
}
\value{
Un dataframe con los anagramas encontrados.
}
\description{
Buscar anagramas
}
