% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_palabra.R
\name{buscar_palabra}
\alias{buscar_palabra}
\title{Buscar una palabra en el diccionario RAE}
\usage{
buscar_palabra(consulta)
}
\arguments{
\item{consulta}{La palabra a buscar.}
}
\value{
Un dataframe con columnas: palabra, resultado, id.
}
\description{
Buscar una palabra en el diccionario RAE
}
\keyword{internal}
