% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandir_categoria}
\alias{expandir_categoria}
\title{Expandir abreviaturas de categoria gramatical}
\usage{
expandir_categoria(abrev)
}
\arguments{
\item{abrev}{La abreviatura (ej. "m.", "tr.").}
}
\value{
La descripcion completa.
}
\description{
Expandir abreviaturas de categoria gramatical
}
\keyword{internal}
