% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extraer_conjugaciones}
\alias{extraer_conjugaciones}
\title{Extraer tabla de conjugaciones}
\usage{
extraer_conjugaciones(contenido_html)
}
\arguments{
\item{contenido_html}{El contenido HTML crudo de la conjugacion.}
}
\value{
Un dataframe con las conjugaciones.
}
\description{
Extraer tabla de conjugaciones
}
\keyword{internal}
