% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extraer_definiciones}
\alias{extraer_definiciones}
\title{Funciones de parseo para la respuesta de la RAE}
\usage{
extraer_definiciones(contenido_html)
}
\arguments{
\item{contenido_html}{El contenido HTML crudo.}
}
\value{
Un dataframe con las definiciones y metadatos.
}
\description{
Este archivo contiene funciones para procesar el HTML retornado por la API.
Extraer definiciones del HTML
}
\keyword{internal}
