% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extraer_etimologia}
\alias{extraer_etimologia}
\title{Extraer etimologia del HTML}
\usage{
extraer_etimologia(contenido_html)
}
\arguments{
\item{contenido_html}{El contenido HTML crudo.}
}
\value{
Un string con la etimologia.
}
\description{
Extraer etimologia del HTML
}
\keyword{internal}
