% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extraer_id_conjugacion}
\alias{extraer_id_conjugacion}
\title{Extraer ID de conjugacion}
\usage{
extraer_id_conjugacion(contenido_html)
}
\arguments{
\item{contenido_html}{El contenido HTML crudo.}
}
\value{
El ID de conjugacion o NA si no existe.
}
\description{
Extraer ID de conjugacion
}
\keyword{internal}
