% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_definicion.R
\name{obtener_definicion}
\alias{obtener_definicion}
\title{Obtener las definiciones de una palabra}
\usage{
obtener_definicion(palabra_consulta = NULL, id = NULL)
}
\arguments{
\item{palabra_consulta}{La palabra a buscar (opcional).}

\item{id}{El ID de la palabra (opcional).}
}
\value{
Un dataframe con las definiciones.
}
\description{
Obtener las definiciones de una palabra
}
\examples{
\donttest{
obtener_definicion("programar")
}
}
