% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_palabra.R
\name{obtener_palabra}
\alias{obtener_palabra}
\title{Obtener informacion completa de una palabra}
\usage{
obtener_palabra(palabra_consulta = NULL, id = NULL, conjugaciones = FALSE)
}
\arguments{
\item{palabra_consulta}{La palabra a buscar (opcional).}

\item{id}{El ID de la palabra (opcional).}

\item{conjugaciones}{Si es TRUE, intenta obtener las conjugaciones (solo para verbos).}
}
\value{
Una lista con la palabra, etimologia, definiciones y conjugaciones (si aplica).
}
\description{
Obtener informacion completa de una palabra
}
