% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_palabra_cruda.R
\name{obtener_palabra_cruda}
\alias{obtener_palabra_cruda}
\title{Obtener la definicion cruda de una palabra por su ID o busqueda}
\usage{
obtener_palabra_cruda(query = NULL, id = NULL)
}
\arguments{
\item{query}{La palabra a buscar si no se provee ID (opcional).}

\item{id}{El ID de la palabra (opcional).}
}
\value{
El contenido HTML de la definicion.
}
\description{
Obtener la definicion cruda de una palabra por su ID o busqueda
}
\keyword{internal}
