% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{palabra_aleatoria}
\alias{palabra_aleatoria}
\title{Obtener una palabra aleatoria}
\usage{
palabra_aleatoria()
}
\value{
Un dataframe con la definicion de una palabra aleatoria.
}
\description{
Obtener una palabra aleatoria
}
