% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{palabra_del_dia}
\alias{palabra_del_dia}
\title{Obtener la palabra del día}
\usage{
palabra_del_dia()
}
\value{
Un dataframe con la palabra del día y su fecha (si está disponible) o contenido HTML crudo.
}
\description{
Obtener la palabra del día
}
