% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{rae_client}
\alias{rae_client}
\title{Cliente RAE interno}
\usage{
rae_client(cadena_consulta)
}
\arguments{
\item{cadena_consulta}{La cadena de consulta de la API (por ejemplo "search?w=hola" o "wotd?callback=json").}
}
\value{
El contenido raw de la respuesta.
}
\description{
Cliente RAE interno
}
\keyword{internal}
